requirements/base.txt
Django==5.0.7
django-environ==0.11.2
psycopg==3.2.3
djangorestframework==3.15.2
dj-rest-auth==6.0.0 